agent-evals-sample/requirements.txt
openai>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
azure-identity>=1.15.0
